Refactor logs in processes
